unit tests from resource_storage.rs moved to tests/unit/filters/resource_storage.rs
